requirements.txt
pillow==11.3.0
Flask==3.1.1
transformers==4.55.2
torch==2.8.0
